backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
firebase-admin
